Fix matmul() 3-D tensor example code
It is necessary to specify an int32 dtype because the code as written
would use int64 dtype by default, and int64 is not currently supported
by matmul().

The "2-D tensor" example code is fine.

+++ tensorflow/python/ops/math_ops.py
@@ -1682,13 +1682,15 @@ def matmul(a,
 
 
   # 3-D tensor `a`
-  a = tf.constant(np.arange(1,13), shape=[2, 2, 3]) => [[[ 1.  2.  3.]
+  a = tf.constant(np.arange(1, 13, dtype=np.int32),
+                  shape=[2, 2, 3])                  => [[[ 1.  2.  3.]
                                                          [ 4.  5.  6.]],
                                                         [[ 7.  8.  9.]
                                                          [10. 11. 12.]]]
 
   # 3-D tensor `b`
-  b = tf.constant(np.arange(13,25), shape=[2, 3, 2]) => [[[13. 14.]
+  b = tf.constant(np.arange(13, 25, dtype=np.int32),
+                  shape=[2, 3, 2])                   => [[[13. 14.]
                                                           [15. 16.]
                                                           [17. 18.]],
                                                          [[19. 20.]
